Add functionality for run tool in root folder
